+++ src/main/java/com/atlauncher/data/LauncherVersion.java
@@ -112,4 +112,13 @@ public class LauncherVersion {
         return String.format(Locale.ENGLISH, "%d.%d.%d.%d %s", this.reserved, this.major, this.minor, this.revision,
                 this.stream);
     }
+
+    public String toStringForUserAgent() {
+        if (this.isReleaseStream()) {
+            return String.format(Locale.ENGLISH, "%d.%d.%d.%d", this.reserved, this.major, this.minor, this.revision);
+        }
+
+        return String.format(Locale.ENGLISH, "%d.%d.%d.%d.%s", this.reserved, this.major, this.minor, this.revision,
+                this.stream);
+    }
 }